Fix gender calculation error and intersex gender chances
The commit changes the gender calculation to avoid an error when
a character is intersex and the gender assignment random int is < 10
which result in the character having no gender variable assigned.

Also, it ensures that the gender assigned to intersex characters is
more evenly distributed as currently an intersex character is male
only if g < 30 but it needs to be > 20 to begin with which means
the chance to be intersex and male is lower than other options.